+++ NEWS.md
@@ -1,6 +1,7 @@
 # pathfindR (development version)
 ## Minor Changes
 - added the `disable_parallel` argument in `active_snw_enrichment_wrapper()` to be able to disable parallel runs via `foreach` 
+- fixed the issue encountered on CentOS where `forech` wasn't loading `pathfindR` (#164)
 
 # pathfindR 2.1.0
 ## Minor Changes